Start tracking with git
At this state, there are three simple command-line-only scripts in this
project. The anki-merge-imports.py script is an old one that I shouldn't
ever need to use again in the future, but I've kept it here in case
there's any useful code in it. The anki-download-student-photos.py
script is the most recent and still reasonably well-working version of
the program that would download the photos directly from my.ucla.
Because of the convoluted way it dealt with my.ucla's authentication,
that project has since been abandoned in favor of the current approach,
which is in anki-pdf-import.py. Much of the code from the former appears
in the latter, roughly intact but vastly reorganized. The most recent
attempt has also been designed to be somewhat modular, to support some
testing that should be done, and a possible future graphical user
interface.

To-do:
    - Test the PDF reader code extensively to make sure it is extracting
      names and student ID numbers correctly, and associating the
      correct photos to each ID number.
    - Add a graphical user interface, instead of or in addition to the
      CLI.
    - Probably split up the latest version into a module containing the
      code that does most of the work, and a simple script that
      implements the "interface" and uses that module. The code is
      already designed to make this separation trivial.