Refactor photoroster module into PhotoRoster class
Put much of the code in the photoroster module (except for the exsting
Student class and the load_existing_students function) into a new class
called PhotoRoster, and modify anki-pdf-import.py (slightly) to work
with the new changes. The new design is very nice.

Also add a little code to keep track of students in existing Anki
collection who already had the tag from the current course, and report
those students who are not in the new photo roster.